hot_post/post/views.py
from django.shortcuts import render
from post.models import Post
from post.serializers import PostSerializer
from rest_framework import viewsets
from rest_framework.parsers import JSONParser
from rest_framework.permissions import IsAuthenticated
# Create your views here.
class PostViewSet(viewsets.ModelViewSet):
    queryset = Post.objects.all()
    serializer_class = PostSerializer
    permission_classes=(IsAuthenticated,)