requirements.txt
Django==3.2
psycopg2>=2.8.6,<2.8.7
djangorestframework==3.13.1
djoser==2.1.0
djangorestframework-jwt==1.11.0
